Deleted old test classes, made new test class
I deleted the EquipmentMonitor and Feeder test classes.
I also created a new Test class EquipmentCheck which is an improved copy
of EquipmentMonitorThread(and deleted this one).
EquipmentCheck was made to be able to handle multiple tanks and only
outputs errors once each time it breaks, so it doesn't flood the MI with
useless notifications.
Still need to incorporate the databases and work on the GUI.