New class EquipmentMonitor
I took the old EquipmentCheck class and put it inside a new class called
EquipmentMonitor, and also deleted the old class so I still only work
with one java file. I added stuff to be able to deal with new entries in
a database.
Work to be done: Add tables to the database with information on
equipment status and have all EquipmentMonitor activity be displayed by
some sort of GUI.