EquipmentMonitor.java and GuiTester.java
Made EquipmentMonitor.java simpler to use by adding new methods to add
and retrieve the list of equipment. Changed how equipment_name and
error_notice are determined, but they will be changed again when I
figure out how to access the database for these values instead.
Also created a new class called GuiTester. It's not a part of the
project yet. It's just a way for me to experiment how I want the module
to look in the end. More work on this is needed.