Deleted new classes created from last commit
I used the copies of the original EquipmentMonitor and GuiTester to
figure out how to integrate the GUI into the Equipment monitor. Once I
figured that out, I did the integration with the original
EquipmentMonitor.java file, so the other .java files became unnecessary.

I still kept the GuiTester.java in case I needed to test other GUI
components, but right now, the only file necessary for the Equipment
Monitor is EquipmentMonitor.java, since that now has everything in it.