Modified EquipmentMonitor.java
Added date and time to the output (both the console and the GUI).
Tried to fix a small GUI bug that doesn't display the text area boxes
correctly. Usually occurs when there are a lot of tanks. If you maximize
the window size, the boxes show up, but when minimized, the boxes appear
way too small. Still not fixed.